requirements.txt
Django>=4.2.17
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1